Migrate to current Roc syntax
- No backpassing
- New record builders
- PNC
- snake_case

The API changes from:

```roc
    |> Pg.Cmd.expectN (
        Pg.Result.succeed {
            name: <- Pg.Result.str "name" |> Pg.Result.apply,
            price: <- Pg.Result.dec "price" |> Pg.Result.apply
        }
    )
```
To:
```roc
|> Pg.Cmd.expect_n(
    { Pg.Result.record_builder <-
        name: Pg.Result.str("name"),
        price: Pg.Result.dec("price"),
    },
)
```

Note `Pg.Result.record_builder` above. `Pg.Cmd.succeed` can still be
used to return non-records like so:

```roc
|> Pg.Cmd.expect_n(
    Pg.Result.succeed(
        |name|
            |age|
                age_str = Num.to_str(age)
                "${name}: ${age_str}",
    )
    |> Pg.Result.with(Pg.Result.str("name"))
    |> Pg.Result.with(Pg.Result.u8("age")),
)
```

It's unclear to me whether this is desirable. However, given that the
name `Pg.Result.succeed` is taken, I opted for
`Pg.Result.record_builder` for the records to make it discoverable and
understandable.

This is a partial migration as I ran out of steam part-way through
`sql-cli` and `examples/store`.